refactor(web): /simplify pass on nav redesign
Four cleanups from a /simplify review of the nav-redesign branch:

* Extract profileInitials() into lib/profile.ts. AppAvatar.vue had
  copy-pasted the same display-name -> initials derivation that
  ProfileCard.vue already had. Both call sites now share one helper.
* Move `justify-self: end` from App.vue's `:deep(.avatar-wrap)`
  override into AppAvatar.vue's own `.avatar-wrap` rule. Drops the
  layering inversion where the parent reached into a child's
  scoped styles to position it; the avatar component owns its
  positioning now.
* Declare --mobile-tab-bar-h once in assets/colors.css instead of
  having `3.75rem` appear in both App.vue's padding-bottom and
  (implicitly) MobileTabBar.vue. Both files reference the var now,
  with MobileTabBar pinning the var as its min-height.
* docs/web-nav.md: collapse the resolved "Open questions" into a
  "Decisions" section, drop "future" from Sign out (now built),
  and link the Account-sub-page backlog item to issue #92.

Skipped from the review (out of scope or unverifiable from this
session): MobileTabBar v-for refactor (needs v-html on SVG, hard
to verify without browser smoke), avatar-circle component
extraction (would refactor ProfileCard layout outside this diff),
DEFAULT_HOME constant (marginal), --bp-mobile breakpoint var
(CSS custom props don't work inside @media at-rules).
